src/RAG-file/requirements.txt
langchain
langchain-community
langchain-openai
qdrant-client
python-dotenv